requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.3.0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
discord==1.7.3
discord.py==1.7.3
idna==3.3
idna-ssl==1.1.0
multidict==5.2.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
numpy==1.22.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.5.2
protobuf==3.19.4
python-slugify==6.1.2
requests==2.27.1
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
